# perf: boost DoMulti/Cache of the cluster by grouping cmds to conns, not slots