related to https://github.com/redis/rueidis/issues/278

The old implementation of `clusterclient.DoMulti()` and `clusterclient.DoMultiCache()` performs poorly when the number of commands passed in is large. That is because those commands will be grouped into their key slots, resulting in many small chunks being parallelly fired. If one's CPU has limited concurrency, then the performance of these functions degrades significantly.

The new implementation grouping commands into their underlying connections, resulting in a minimum number of chunks to be parallelly fired. Even with limited concurrency, this implementation will have a better performance.

Before:
```
Benchmark
Benchmark/GET_DoMulti
Benchmark/GET_DoMulti-10         	     144	   8138918 ns/op	  484360 B/op	    5015 allocs/op
Benchmark/GET_DoMultiCache
Benchmark/GET_DoMultiCache-10    	    6067	    187792 ns/op	  440528 B/op	    4926 allocs/op
PASS
```


After:
```
Benchmark
Benchmark/GET_DoMulti
Benchmark/GET_DoMulti-10         	    2194	    491935 ns/op	  363532 B/op	    1100 allocs/op
Benchmark/GET_DoMultiCache
Benchmark/GET_DoMultiCache-10    	   13555	     88418 ns/op	  349316 B/op	     104 allocs/op
PASS
```