@smf8 2023-06-18
Hey @rueian, the changes looks great to me. There is one minor improvement, the implementation of `mux.DoMultiCache` uses the `ParallelKeys` util function. Having only 1 CPU core will limit the client to execute only 1 command at a time which I believe is limiting. Although it can be modified with `GOMAXPROCS`, but changing this env for this single purpose might not be a good idea. 

Can we increase the lower bound of `concurrency` in `closeThenParallel`? for my usage, increasing it to `runtime.GOMAXPROCS(0) * 4` resulted in around 20% improvement in benchmarks But I think it can increase even more for high latency networks and slow commands.